enhancement_workers/demo-tesseract-lambda-python/requirements.txt
-i https://pypi.org/simple
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
jellyfish==1.0.1; python_version >= '3.7'
langdetect==1.0.9
networkx==3.1; python_version >= '3.8'
numpy==1.26.1; python_version < '3.13' and python_version >= '3.9'
packaging==23.2; python_version >= '3.7'
pillow==10.1.0; python_version >= '3.8'
pytesseract==0.3.9
regex==2023.10.3; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
segtok==1.5.11
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabulate==0.9.0; python_version >= '3.7'
urllib3==2.0.7; python_version >= '3.7'
yake==0.4.8
